pyproject.toml
[project]
name = "snowid"
version = "0.0.1"
description = "Simple 2D game engine built on top of PySDL2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySDL2",
    "IPython",
    "pymunk",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "sdl2.*",
    "pymunk.*",
    "IPython.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["project/font.py"]  # Contains binary data

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "TID252"]
# Ruff is compatible with black by default, no need to ignore W503

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.lint.isort]
known-first-party = ["project"]

[tool.pytest.ini_options]
testpaths = ["project/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=project",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require external dependencies)",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["project"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]


